PYTHON-5784 Cover threaded PeriodicExecutor exception path via synchro
Consolidate the target-exception test into the synchro'd source so both
AsyncPeriodicExecutor and the threaded PeriodicExecutor are covered from a
single source, and delete the async-only test file.

Why this reverses the async-only split from 33409591: that commit carved the
exception test out into test_async_periodic_executor.py to dodge the noisy
traceback an uncaught exception prints via threading.excepthook on the
background thread. But the two _run loops handle the exception differently --
the threaded version additionally takes the lock and sets _thread_will_exit,
which feeds the join-on-reopen path in open(). The async test therefore did
not cover that sync-only branch, leaving a real gap (also flagged by Copilot).

Rather than maintain two near-identical hand-written tests (which cuts against
the synchro direction this PR has otherwise followed), the test now lives in
the async source with an `if _IS_SYNC:` branch that swaps threading.excepthook
to a no-op for the threaded case. Verified the generated sync test passes
15/15 with no flakiness or stderr noise, which confirms the path is testable
and there was no real reason to keep it async-only.
